Se refactorizon los modulos y se corrigieron errores en los controladores

+++ app/Repositories/CustomerRepository.php
@@ -7,15 +7,26 @@ use App\User;
 
 class CustomerRepository
 {
-    public function updateCustomer(Request $request)
+    public function updateCustomer(Request $request, $id)
     {
-        User::where('id', $request->id)->update($request->except(['password_confirmation']));
+        $user = User::find($id);
+
+        if(is_null($user)) {
+            return 422;
+        }
+
+        User::where('id', $user->id)->update($request->except(['password_confirmation']));
         return 200;
     }
 
     public function deleteCustomer($id)
     {
-        $user = User::findOrFail($id);
+        $user = User::find($id);
+
+        if(is_null($user)) {
+            return 422;
+        }
+
         $user->delete();
         return 204;
     }
